created diffrent files for the test cases
