+++ src/modules/arrayAnalysis.js
