+++ src/modules/calculator.js
@@ -1,0 +1,22 @@
+const sum = (a, b) => {
+  return a + b;
+};
+
+const subtract = (a, b) => {
+  return a - b;
+};
+
+const divide = (a, b) => {
+  return a / b;
+};
+
+const multiply = (a, b) => {
+  return a * b;
+};
+
+module.exports = {
+  sum,
+  subtract,
+  divide,
+  multiply,
+};
